Move some code to signatures.py
